Moved model classes so similar types are together
- Moved the model classs so all types are together
