+++ app/Models/ItemType/Summary/AllocatedExpense.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace App\Models\Summary\ItemType;
+namespace App\Models\ItemType\Summary;
 
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Model;
