+++ app/Models/ItemType/Summary/SimpleExpense.php
@@ -1,11 +1,10 @@
 <?php
 declare(strict_types=1);
 
-namespace App\Models\Summary\ItemType;
+namespace App\Models\ItemType\Summary;
 
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Query\Builder as QueryBuilder;
-use Illuminate\Support\Facades\DB;
 
 /**
  * @mixin QueryBuilder
